refactor: centralize location types
