+++ functions/weather/weatherTypes.ts
@@ -1,19 +1,4 @@
-export interface StreetAddress {
-	city?: string;
-	state?: string;
-	country?: string;
-}
-
-export interface ZipCode {
-	zipCode?: string;
-}
-
-export interface GEOCODE {
-	lat?: number;
-	lon?: number;
-}
-
-export interface LocationInput extends StreetAddress, ZipCode, GEOCODE {}
+import { LocationInput } from '../../types/location';
 
 export interface WeatherRequest extends LocationInput {}
